Fixed multiple submission bug
If a student had the same choice open in multiple tabs/windows/browsers,
it was possible to submit multiple times and have multiple (confirmed
and unarchived) results in the database. This was because the second
save would have been detected as a new record, so existing unarchived
records would not have been archived.