Reload when try to changes selection that has a more recent submission
If user tries to change a choice, but they have made a more recent
submission (in another tab/window/browser), a dialog will pop up that
forces them to reload the page. They are also now given the opion to
reload the page (to view the submitted choice) if they try to save from
a page that was loaded before a more recent submission.